+++ desafio-spring/src/main/java/com/example/desafiospring/dto/response/SellerResponseDTO.java
@@ -7,31 +7,31 @@ import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class SellerResponseDTO extends UserResponseDTO {
-    private Long followers_count;
+    private Integer followers_count;
     private List<UserResponseDTO> followers;
 
     public SellerResponseDTO() {
     }
 
-    public SellerResponseDTO(Long followers_count) {
+    public SellerResponseDTO(Integer followers_count) {
         this.followers_count = followers_count;
     }
 
-    public SellerResponseDTO(Long userId, String userName, Long followers_count) {
+    public SellerResponseDTO(Integer userId, String userName, Integer followers_count) {
         super(userId, userName);
         this.followers_count = followers_count;
     }
 
-    public SellerResponseDTO(Long userId, String userName, List<UserResponseDTO> followers) {
+    public SellerResponseDTO(Integer userId, String userName, List<UserResponseDTO> followers) {
         super(userId, userName);
         this.followers = followers;
     }
 
-    public Long getFollowers_count() {
+    public Integer getFollowers_count() {
         return followers_count;
     }
 
-    public void setFollowers_count(Long followers_count) {
+    public void setFollowers_count(Integer followers_count) {
         this.followers_count = followers_count;
     }
