+++ desafio-spring/src/main/java/com/example/desafiospring/repositories/IProductRepository.java
@@ -3,6 +3,6 @@ package com.example.desafiospring.repositories;
 import com.example.desafiospring.exceptions.ProductException;
 import com.example.desafiospring.models.Product;
 
-public interface IProductRepository extends IRepository<Product, Long, ProductException> {
-    Long addProduct(Product product);
+public interface IProductRepository extends IRepository<Product, Integer, ProductException> {
+    Integer addProduct(Product product);
 }
