+++ desafio-spring/src/main/java/com/example/desafiospring/repositories/ISellerRepository.java
@@ -5,12 +5,13 @@ import com.example.desafiospring.models.Seller;
 
 import java.util.List;
 
-public interface ISellerRepository extends IRepository<Seller, Long, SellerException> {
-    Seller addFollower(Long userId, Long userIdToFollow) throws SellerException;
+public interface ISellerRepository extends IRepository<Seller, Integer, SellerException> {
+    
+    Seller addFollower(Integer userId, Integer userIdToFollow) throws SellerException;
 
-    Seller removeFollower(Long userId, Long userIdToFollow) throws SellerException;
+    Seller removeFollower(Integer userId, Integer userIdToFollow) throws SellerException;
 
-    List<Seller> findByFollowerUserId(Long userId);
+    List<Seller> findByFollowerUserId(Integer userId);
 
-    Boolean checkIfSellerExistsById(Long sellerId) throws SellerException;
+    Boolean checkIfSellerExistsById(Integer sellerId) throws SellerException;
 }
