+++ desafio-spring/src/main/java/com/example/desafiospring/repositories/IUserRepository.java
@@ -3,6 +3,6 @@ package com.example.desafiospring.repositories;
 import com.example.desafiospring.exceptions.UserException;
 import com.example.desafiospring.models.User;
 
-public interface IUserRepository extends IRepository<User, Long, UserException> {
-    Boolean checkIfUserExistsById(Long userId) throws UserException;
+public interface IUserRepository extends IRepository<User, Integer, UserException> {
+    Boolean checkIfUserExistsById(Integer userId) throws UserException;
 }
